attr element and new damage calculing

+++ src/creatures/combat/combat.cpp
@@ -1614,7 +1614,7 @@ void ValueCallback::getMinMaxValues(const std::shared_ptr<Player> &player, Comba
 	LuaScriptInterface::pushUserdata<Player>(L, player);
 	LuaScriptInterface::setMetatable(L, -1, "Player");
 
-	int16_t elementAttack = 0; // To calculate elemental damage after executing spell script and get real damage.
+	int32_t elementAttack = 0; // To calculate elemental damage after executing spell script and get real damage.
 	int32_t attackValue = 7; // default start attack value
 	int parameters = 1;
 	bool shouldCalculateSecondaryDamage = false;
